+++ cookbooks/aws-parallelcluster-platform/attributes/platform.rb
@@ -16,11 +16,14 @@ default['cluster']['enroot']['persistent_dir'] = '/var/enroot'
 
 # NVidia
 default['cluster']['nvidia']['enabled'] = 'no'
-default['cluster']['nvidia']['driver_version'] = '570.86.15'
+default['cluster']['nvidia']['driver_version'] = '570.172.08'
 default['cluster']['nvidia']['dcgm_version'] = '3.3.6'
 if platform?('amazon') && node['platform_version'] == "2"
   default['cluster']['nvidia']['driver_version'] = '550.127.08'
 end
+
+# nvidia-imex
+default['cluster']['nvidia']['imex']['shared_dir'] = "#{node['cluster']['shared_dir']}/nvidia-imex"
 
 # DCV
 default['cluster']['dcv']['authenticator']['user'] = "dcvextauth"